requirements.txt
alembic==1.15.1
aniso8601==10.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dropbox==12.0.2
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
imagekitio==4.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.7
oauthlib==3.2.2
openpyxl==3.1.5
ply==3.11
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.38
stone==3.3.1
tenacity==9.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
